% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CondensedDatabase}
\alias{CondensedDatabase}
\title{Condensed Database.}
\format{
A data frame with a total of 10 variables.

Taxonomic variables (3): Species, Family; Butterfly (1 for Papilionoidea Superfamily)

Distribution Variables (3): Region, Realm, Continent;

Introduction Variables (2): First Observation, Intentional Release;

Reference Data (2): Latest Reference (in the Complete database), Latest Reference Year
}
\usage{
CondensedDatabase
}
\description{
Condensed Database to one Species per Region.
}
\keyword{datasets}
