% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CondensedGNOLEP}
\alias{CondensedGNOLEP}
\title{Condensed GNOLEP}
\format{
A data frame with 3857 entries and a total of 13 variables.

Taxonomic variables (1): Accepted Species according to GBIF;

Distribution Variables (3): Region, Continent, Realm;

Introduction Variables (7): Cryptogenic, Intentional Release, Introduced, Dispersal, Established, Eradicated, First Record Year;

Reference Data (2): Latest Reference (in the Complete database), Latest Reference Year


\describe{
  \item{AcceptedSpecies}{character; species name accepted by GBIF}
  \item{AreaName}{character; record region}
  \item{Continent}{character; record continent}
  \item{Realm}{character; record biogeographic realm}
  \item{First_Record}{integer; oldest first record}
  \item{Cryptogenic}{binary; Cryptogenic Origin}
  \item{Dispersal}{binary; Arrival by dispersal}
  \item{Eradicated}{binary; Species previously established but currently eradicated}
  \item{Established}{binary; record of established population}
  \item{IntentionalRelease}{binary; record of intentional introduction (biocontrol included)}
  \item{Introduced}{binary; record of arrival related to human activity}
  \item{Latest_Reference}{character; most recent reference on the database for the species-region pair}
  \item{Latest_Reference_Year}{integer; reference year}
}
}
\usage{
CondensedGNOLEP
}
\description{
Condensed Database to one Species per Region.
}
\keyword{datasets}
