% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DistributionData}
\alias{DistributionData}
\title{Raw distribution CSV}
\format{
A data frame with ... (describe columns)
}
\usage{
DistributionData
}
\description{
Original CSV read into R as `DistributionData`.
}
\keyword{datasets}
