% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GNOLEP}
\alias{GNOLEP}
\title{GNOLEP - Worldwide Database of non-Native Lepidoptera}
\format{
A data frame with 11916 records and 18 variables:

Database Variables (1): RecordID;

Introduction Variables (7): Cryptogenic, Intentional Release, Introduced, Dispersal, Established, Eradicated, First Record Year;

Taxonomic variables (3): Species as in the Original Record, Species as Accepted in GBIF, Genus, Family;

Distribution Variables (4): Region, Country, Continent, Biogeographic Realm;

Reference Data (2): Reference Citation, Reference Year;

\describe{
  \item{GNOLEP.recordID}{integer; unique record id}
  \item{GNOLEP.cryptogenic}{binary; Cryptogenic Origin}
  \item{GNOLEP.intentionalRelease}{binary; record of intentional introduction (biocontrol included)}
  \item{GNOLEP.introduced}{binary; record of human mediated introduction }
  \item{GNOLEP.dispersal}{binary; record of species arriving by dispersal }
  \item{GNOLEP.established}{binary; record of established population}
  \item{GNOLEP.eradicated}{binary; record of eradication of previously established population}
  \item{dwc.year}{integer; year of first record}
  \item{dwc.scientificName}{character; species name as in the original source}
  \item{GNOLEP.acceptedSpecies}{character; species name accepted by GBIF}
  \item{dwc.genus}{character; species taxonomic genus}
  \item{dwc.family}{character; species taxonomic family}
  \item{dwc.verbatimLocality}{character; record region}
  \item{dwc.country}{character; record country}
  \item{dwc.continent}{character; record continent}
  \item{GNOLEP.realm}{character; record biogeographic realm}
  \item{dwc.associatedReferences}{character; original record reference}
  \item{GNOLEP.referenceYear}{integer; reference year}
}
}
\source{
GNOLEP database (check package source reference)
}
\usage{
GNOLEP
}
\description{
A processed dataset containing the distribution `Records` joined with taxonomy, regions,
realms and references. This dataset is produced directly from the GNOLEP SQLite database
(see `data-raw/data-preparation.R`).
}
\keyword{datasets}
