% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_GNOLEP.R
\name{GNOLEP_extract}
\alias{GNOLEP_extract}
\title{Import older versions of GNOLEP}
\usage{
GNOLEP_extract(table, version = "", destination_file = NULL)
}
\value{
Import all database tables at once.
}
\description{
Import older versions of GNOLEP
}
\examples{

# GNOLEP_extract(NativesGNOLEP, version = 0)
# GNOLEP_extract(ReferencesGNOLEP, version = 0)
# GNOLEP_extract(GNOLEP, version = 0)
# GNOLEP_extract(CondensedGNOLEP, version = 0)



}
