% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_GNOLEP.R
\name{individual_GNOLEP}
\alias{individual_GNOLEP}
\title{All Individual GNOLEP tables imports}
\usage{
individual_GNOLEP()
}
\value{
Import all individual database tables at once.
}
\description{
All Individual GNOLEP tables imports
}
\examples{
# full_GNOLEP()

}
