% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_WDnNL.R
\name{individual_WDnNL}
\alias{individual_WDnNL}
\title{All Individual WDnNL tables imports}
\usage{
individual_WDnNL()
}
\value{
Import all individual database tables at once.
}
\description{
All Individual WDnNL tables imports
}
\examples{
# full_WDnNL()

}
