% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NativeDistribution}
\alias{NativeDistribution}
\title{NativeDistribution table}
\format{
A data frame...
}
\usage{
NativeDistribution
}
\description{
NativeDistribution table
}
\keyword{datasets}
