% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NativesGNOLEP}
\alias{NativesGNOLEP}
\title{Native Distribution Table}
\format{
A data frame with
}
\usage{
NativesGNOLEP
}
\description{
Table with Native Distribution per Species
}
\keyword{datasets}
