% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_GNOLEP.R
\name{print.summary.CondensedGNOLEP}
\alias{print.summary.CondensedGNOLEP}
\title{Print method for CondensedGNOLEP summary}
\usage{
\method{print}{summary.CondensedGNOLEP}(x, ...)
}
\arguments{
\item{x}{A summary.CondensedGNOLEP object}

\item{...}{Not Used}
}
\value{
A printed summary of the dataframe
}
\description{
Print method for CondensedGNOLEP summary
}
