% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_WDnNL.R
\name{print.summary.CondensedWDnNL}
\alias{print.summary.CondensedWDnNL}
\title{Print method for CondensedWDnNL summary}
\usage{
\method{print}{summary.CondensedWDnNL}(x, ...)
}
\arguments{
\item{x}{A summary.CondensedWDnNL object}

\item{...}{Not Used}
}
\value{
A printed summary of the dataframe
}
\description{
Print method for CondensedWDnNL summary
}
