% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_GNOLEP.R
\name{print.summary.GNOLEP}
\alias{print.summary.GNOLEP}
\title{Print method for GNOLEP summary}
\usage{
\method{print}{summary.GNOLEP}(x, ...)
}
\arguments{
\item{x}{A summary.GNOLEP object}

\item{...}{Not Used}
}
\value{
A printed summary of the GNOLEP dataframe
}
\description{
Print method for GNOLEP summary
}
