% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_GNOLEP.R
\name{summary.CondensedGNOLEP}
\alias{summary.CondensedGNOLEP}
\title{Summary of the CondensedGNOLEP as a summary.CondensedGNOLEP object}
\usage{
\method{summary}{CondensedGNOLEP}(
  object,
  include_non_established = TRUE,
  include_absent_establishment_data = TRUE,
  include_intentional = TRUE,
  include_absent_intentional_data = TRUE,
  include_cryptogenic = TRUE,
  include_absent_cryptogenic_data = TRUE,
  include_non_Introduced = TRUE,
  include_absent_Introduced_data = TRUE,
  include_eradicated_data = TRUE,
  include_eradication_records = TRUE,
  include_records_pre_eradication = TRUE,
  continent_list = NULL,
  realm_list = NULL
)
}
\arguments{
\item{object}{A CondensedGNOLEP dataframe}

\item{include_non_established}{TRUE or FALSE - define if keep only established species or all}

\item{include_absent_establishment_data}{TRUE or FALSE - define if keep records with NA's on establishement data}

\item{include_intentional}{TRUE or FALSE - define if keep records from intentional release}

\item{include_absent_intentional_data}{TRUE or FALSE - define if keep records with NA's on intentional release}

\item{include_cryptogenic}{TRUE or FALSE - define if keep records with cryptogenic origin}

\item{include_absent_cryptogenic_data}{TRUE or FALSE - define if keep records with NA's on cryptogenic origin}

\item{include_non_Introduced}{TRUE or FALSE - define if keep species that arrived through natural dispersal}

\item{include_absent_Introduced_data}{TRUE or FALSE - define if keep records with NA's on introduction origin}

\item{include_eradicated_data}{TRUE or FALSE - define if keep records with NA's on erradicated}

\item{include_eradication_records}{TRUE or FALSE - define if keep records with successful erradication}

\item{include_records_pre_eradication}{TRUE or FALSE - define if keep records of introductions after erradications}

\item{continent_list}{A list of Continents}

\item{realm_list}{A list of Biogeographic Realms}
}
\value{
A printed summary of the CondensedGNOLEP
}
\description{
Summary of the CondensedGNOLEP as a summary.CondensedGNOLEP object
}
