% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxonomyGNOLEP}
\alias{TaxonomyGNOLEP}
\title{Taxonomy table}
\format{
A data frame with the Taxonomy table from the GNOLEP.

\describe{
  \item{GNOLEP.speciesID}{character; species unique identifier}
  \item{GNOLEP.acceptedSpeciesID}{character; species unique identifier related to the current species name as it is accepted by GBIF}
  \item{dwc.family}{character; family name}
  \item{dwc.genus}{character; genus name}
  \item{dwc.scientificName}{character; species name}
}
}
\usage{
TaxonomyGNOLEP
}
\description{
Taxonomy table
}
\keyword{datasets}
