% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxonomyWDnNL}
\alias{TaxonomyWDnNL}
\title{Taxonomy table}
\format{
A data frame with the Taxonomy table from the WDnNL.

\describe{
  \item{SpeciesID}{character; species unique identifier}
  \item{AcceptedSpeciesID}{character; species unique identifier related to the current species name as it is accepted by GBIF}
  \item{Family}{character; family name}
  \item{Genus}{character; genus name}
  \item{Species}{character; species name}
}
}
\usage{
TaxonomyWDnNL
}
\description{
Taxonomy table
}
\keyword{datasets}
