% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WDnNL}
\alias{WDnNL}
\title{WDnNL - Worldwide Database of non-Native Lepidoptera}
\format{
A data frame with 11916 records and 18 variables:

Database Variables (1): RecordID;

Introduction Variables (7): Cryptogenic, Intentional Release, Introduced, Dispersal, Established, Eradicated, First Record Year;

Taxonomic variables (3): Species as in the Original Record, Species as Accepted in GBIF, Genus, Family;

Distribution Variables (4): Region, Country, Continent, Biogeographic Realm;

Reference Data (2): Reference Citation, Reference Year;

\describe{
  \item{RecordID}{integer; unique record id}
  \item{Cryptogenic}{binary; Cryptogenic Origin}
  \item{IntentionalRelease}{binary; record of intentional introduction (biocontrol included)}
  \item{Introduced}{binary; record of human mediated introduction }
  \item{Dispersal}{binary; record of species arriving by dispersal }
  \item{Established}{binary; record of established population}
  \item{Eradicated}{binary; record of eradication of previously established population}
  \item{Year}{integer; year of first record}
  \item{Species}{character; species name as in the original source}
  \item{AcceptedSpecies}{character; species name accepted by GBIF}
  \item{Genus}{character; species taxonomic genus}
  \item{Family}{character; species taxonomic family}
  \item{AreaName}{character; record region}
  \item{Country}{character; record country}
  \item{Continent}{character; record continent}
  \item{Realm}{character; record biogeographic realm}
  \item{BibliographicReference}{character; original record reference}
  \item{ReferenceYear}{integer; reference year}
}
}
\source{
WDNnL database (check package source reference)
}
\usage{
WDnNL
}
\description{
A processed dataset containing the distribution `Records` joined with taxonomy, regions,
realms and references. This dataset is produced directly from the WDNnL SQLite database
(see `data-raw/data-preparation.R`).
}
\keyword{datasets}
