% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/full_WDnNL.R
\name{WDnNL_extract}
\alias{WDnNL_extract}
\title{Import older versions of WDnNL}
\usage{
WDnNL_extract(table, version = "", destination_file = NULL)
}
\value{
Import all database tables at once.
}
\description{
Import older versions of WDnNL
}
\examples{

# WDnNL_extract(NativesWDnNL, version = 0)
# WDnNL_extract(ReferencesWDnNL, version = 0)
# WDnNL_extract(WDnNL, version = 0)
# WDnNL_extract(CondensedWDnNL, version = 0)



}
