% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_WDnNL.R
\name{plot.WDnNL}
\alias{plot.WDnNL}
\title{Plot Data Summaries}
\usage{
\method{plot}{WDnNL}(
  object,
  OverTime = TRUE,
  ContinentalPlot = FALSE,
  RealmPlot = FALSE,
  include_non_established = TRUE,
  include_absent_establishment_data = TRUE,
  include_intentional = TRUE,
  include_absent_intentional_data = TRUE,
  include_cryptogenic = TRUE,
  include_absent_cryptogenic_data = TRUE,
  include_non_Introduced = TRUE,
  include_absent_Introduced_data = TRUE,
  include_eradicated_data = TRUE,
  include_eradication_records = TRUE,
  include_records_pre_eradication = TRUE,
  continent_list = NULL,
  family_list = NULL,
  realm_list = NULL
)
}
\arguments{
\item{object}{A WDnNL object dataframe}
}
\value{
Summary plots for the subset of WDnNL selected.
}
\description{
Plot Data Summaries
}
\examples{
# plot(WDnNL)

# plot(WDnNL,
       OverTime = FALSE,
       ContinentalPlot = TRUE,
       include_non_established = FALSE,
          include_absent_establishment_data = FALSE,
          include_cryptogenic = FALSE,
          include_non_Introduced = FALSE,
          family_list = c("Noctuidae", "Gelechiidae", "Tortricidae"))

}
