% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_WDnNL.R
\name{print.summary.CondensedWDnNL}
\alias{print.summary.CondensedWDnNL}
\title{Title}
\usage{
\method{print}{summary.CondensedWDnNL}(x, ...)
}
\arguments{
\item{x}{A summary.WDnNL object}
}
\value{
A printed summary of the dataframe
}
\description{
Title
}
