% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary_WDnNL.R
\name{summary.WDnNL}
\alias{summary.WDnNL}
\title{Title}
\usage{
\method{summary}{WDnNL}(
  object,
  include_non_established = TRUE,
  include_absent_establishment_data = TRUE,
  include_intentional = TRUE,
  include_absent_intentional_data = TRUE,
  include_cryptogenic = TRUE,
  include_absent_cryptogenic_data = TRUE,
  include_non_Introduced = TRUE,
  include_absent_Introduced_data = TRUE,
  include_eradicated_data = TRUE,
  include_eradication_records = TRUE,
  include_records_pre_eradication = TRUE,
  family_list = NULL,
  country_list = NULL,
  continent_list = NULL,
  realm_list = NULL
)
}
\arguments{
\item{realm_list}{}
}
\description{
Title
}
