% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_extract.R
\name{version_extract}
\alias{version_extract}
\title{Table Extraction from previous database versions.}
\usage{
version_extract(version = "", table)
}
\arguments{
\item{version}{The database version to extract}

\item{table}{The table to extract from the version}
}
\value{
The table from the specified version as a dataframe
}
\description{
Table Extraction from previous database versions.
}
\examples{
version_extract(,"NativeDatabase")
}
